Mocking pdf service
